+++ dbus-mqtt-opendtu/restart.sh
@@ -11,7 +11,8 @@ if [ -n "$pid" ]; then
     pkill -f "python $SCRIPT_DIR/$SERVICE_NAME.py" > /dev/null 2>&1
     echo "done."
 else
-    echo "driver is not running!"
+    svc -u /service/$SERVICE_NAME
+    echo "started."
 fi
 
 echo